requirements.txt
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langgraph-api
langgraph-cli
langsmith
langchain-community
langchain-core
langchain-openai
langchain-anthropic
langchain-tavily
langchain-google-community[gmail,calendar]

tavily
markdownify
deepagents
trustcall
dotenv
e2b-code-interpreter
